+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-## v0.1.2 [unreleased]
+## v0.1.3 [unreleased]
+
+### Release Notes
+
+### Bugfixes
+
+
+## v0.1.2 [2017-05-02]
 
 ### Release Notes